Use index.js files for exporting modules, only. Differentiate between Containers and Views in file names.
