+++ src/core/components/Home/HomeView.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+function HomeView() {
+  return <h1>Kibana home</h1>
+}
+
+export default HomeView
