+++ src/core/components/Home/index.js
@@ -1,7 +1,3 @@
-import React from 'react'
-
-function Home() {
-  return <h1>Kibana home</h1>
-}
-
-export default Home
+export {
+  default as HomeContainer,
+} from './HomeContainer'
