+++ src/core/components/index.js
@@ -1,0 +1,3 @@
+export {
+  default as AppContainer
+} from './AppContainer';
